requirements.txt
PyYAML==6.0.2
pydantic==2.9.2
strictyaml==1.7.3
hydra-core==1.3.2
